openhours cache

+++ server/src/resources/openhour/resolvers/openhour.resolver.ts
@@ -6,7 +6,7 @@ import {
   WhereOpenHour,
 } from "../../../models/openhour.model";
 import { OpenHourService } from "../services/openhour.service";
-import { ForbiddenException, UseGuards } from "@nestjs/common";
+import { UseGuards } from "@nestjs/common";
 import { JwtAuthGuard } from "../../../auth/guards/jwt.guard";
 import { RoleGuard } from "../../../auth/guards/role.guard";
 import { RESTAURANT, WAITER } from "../../../role";
@@ -16,14 +16,30 @@ import { User } from "../../../auth/decorators/user.decorator";
 import { JwtPayload } from "../../../interfaces/jwt.interface";
 import { IdIntercept } from "../../../auth/guards/id.guard";
 import { PermissionDeniedException } from "../../../error";
+import { CacheService } from "../../../cache/services/cache.service";
 
 @Resolver((of) => OpenHour)
 export class OpenHourResolver {
-  constructor(private readonly openHourService: OpenHourService) {}
+  constructor(
+    private readonly openHourService: OpenHourService,
+    private readonly cacheService: CacheService
+  ) {}
+
+  private cachePrefix(restaurantId: number) {
+    return `openhours:${restaurantId}`;
+  }
+
+  private clearCache(restaurantId: number) {
+    this.cacheService.del(this.cachePrefix(restaurantId));
+  }
+
   @Mutation(() => OpenHour, { name: "createOpenHour" })
   @UseGuards(JwtAuthGuard, RoleGuard(RESTAURANT))
   async create(@Args("data") data: CreateOpenHour, @User() { id }: JwtPayload) {
     const openHour = await this.openHourService.create(data, id);
+
+    this.clearCache(id);
+
     return openHour;
   }
 
@@ -33,8 +49,11 @@ export class OpenHourResolver {
     @Args("data", { type: () => [CreateOpenHour] }) data: CreateOpenHour[],
     @User() { id }: JwtPayload
   ) {
-    await this.openHourService.createMany(data, id);
-    return { message: "success" };
+    const openHours = await this.openHourService.createMany(data, id);
+
+    this.clearCache(id);
+
+    return openHours;
   }
 
   @Mutation(() => OpenHour, { name: "updateOpenHour" })
@@ -52,14 +71,38 @@ export class OpenHourResolver {
 
   @Mutation(() => Success, { name: "deleteOpenHour" })
   @UseGuards(JwtAuthGuard, RoleGuard(RESTAURANT))
-  async delete(@Args("where") where: WhereOpenHour) {
+  async delete(
+    @Args("where") where: WhereOpenHour,
+    @User() { id }: JwtPayload
+  ) {
+    const openHour = await this.openHourService.find(where);
+    if (openHour.restaurantId != id) throw new PermissionDeniedException();
+
     await this.openHourService.delete(where);
+
+    this.clearCache(id);
+
     return { message: "success" };
   }
 
   @Query(() => [OpenHour], { name: "openHours" })
   @UseGuards(JwtAuthGuard, RoleGuard(RESTAURANT, WAITER), IdIntercept)
-  list(@RID() restaurantId: number) {
-    return this.openHourService.list(restaurantId);
+  async list(@RID() restaurantId: number) {
+    const cached = await this.cacheService.get({
+      key: this.cachePrefix(restaurantId),
+      json: true,
+    });
+
+    if (cached) return cached;
+
+    const openHours = await this.openHourService.list(restaurantId);
+
+    this.cacheService.set({
+      key: this.cachePrefix(restaurantId),
+      value: JSON.stringify(openHours),
+      ttl: 120_000,
+    });
+
+    return openHours;
   }
 }
